Rearrange lapack/cusolver headers
